Show more info in API
Show more information in the plants API, so you
don't have to query for every plant and waste a
lot of astrobotany's server resources.

This could lead to being a privacy issue, as
having all the information in one place means
someone can easily get the data, when they
would have to implement their own parsers and
other things to get this data.

If this idea isn't wanted because the API would
like to be kept the same, I don't mind it having
under another API endpoint.